refactor code in UserInterface class
